backend/requirements.txt
alembic==1.14.1
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic-settings==2.8.1
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
feedparser==6.0.11
pandas==2.2.3
numpy==2.2.4
httpx==0.28.1
openai==1.68.2
python-dotenv==1.0.1
pytest==8.3.5
